Wind River Distro: enable container and image builds with WR
This requires a base container created by a LTS23 source build (for
now). See the reference build config required to produce this
image. It is imported using

 podman import -c CMD=/usr/bin/bash tmp-glibc/deploy/images/x86-64/wr-app-container-x86-64.tar.bz2 basic-container:latest

This then requires a LTS23 source build (for now) containing all the
required packages for the COSA container to be complete. The RPMs are
exposed as a package feed.

 ruby -run -ehttpd tmp-glibc/deploy/rpm -p5555

Once the package feed is ready the COSA container can be built and
tagged.

 podman build .
 podman tag <container_id> quay.io/coreos-assembler/coreos-assembler:latest

The COSA container can then be used per the instructions at
https://coreos.github.io/coreos-assembler/building-fcos/#define-a-bash-alias-to-run-cosa

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>